+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add documentation on how to update a deployment of the serverless demo.
+- Update background blur options to allow for skipping frames to improve CPU utilization.
 
 ### Removed
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Change the serverless demo's `deploy.js` script to rebuild the demo on each run. This should remove a manual step of rebuilding the demo.
 - Disable NScale resolution scaling for Android device due to Android H.264 encoding.
+- Minor clean up of code in `BackgroundBlurProcessorProvided` class.
 
 ## [2.21.0] - 2021-11-01